refactor(frontend): move mainskills into subfolders
Group mainskills by their base skill (e.g. Lunar Blessing, Berry Juice, and Energy for Everyone go together) into subfolders. The existing folder has a ton of files in it, making it harder to navigate and find specific files.
