+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/berry-burst/berry-burst-details.vue
