+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/berry-burst/disguise-berry-burst-details.vue
