+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/charge-energy-s/moonlight-charge-energy-s-details.test.ts
