+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/charge-strength-m/bad-dreams-charge-strength-m-details.test.ts
