+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/charge-strength-m/charge-strength-m-details.vue
