+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/charge-strength-s/charge-strength-s-range-details.test.ts
