+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/charge-strength-s/stockpile-charge-strength-s-details.test.ts
