+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/cooking-assist-s/bulk-up-cooking-assist-s-details.test.ts
