+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/cooking-power-up-s/minus-cooking-power-up-s-details.vue
