+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/dream-shard-magnet-s/dream-shard-magnet-s-details.vue
