+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/energizing-cheer-s/heal-pulse-energizing-cheer-s-details.test.ts
