+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/energy-for-everyone-s/energy-for-everyone-s-details.test.ts
