+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/extra-helpful-s/extra-helpful-s-details.test.ts
