+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/helper-boost/helper-boost-details.test.ts
