+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/ingredient-draw-s/hyper-cutter-ingredient-draw-s-details.test.ts
