+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/ingredient-draw-s/ingredient-draw-s-details.vue
