+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/ingredient-draw-s/super-luck-ingredient-draw-s-details.vue
