+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/ingredient-magnet-s/ingredient-magnet-s-details.test.ts
