+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/metronome/metronome-details.vue
