+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/skill-copy/transform-skill-copy-details.vue
