+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/tasty-chance-s/tasty-chance-s-details.test.ts
