+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill.vue
@@ -14,7 +14,7 @@
 import { mainskillImage } from '@/services/utils/image-utils'
 import { useTeamStore } from '@/stores/team/team-store'
 import type { MemberProductionExt } from '@/types/member/instanced'
-import { MathUtils } from 'sleepapi-common'
+import { MathUtils, ModifiedMainskill } from 'sleepapi-common'
 import { defineAsyncComponent, defineComponent, type PropType } from 'vue'
 
 export default defineComponent({
@@ -31,18 +31,28 @@ export default defineComponent({
   },
   computed: {
     skillComponent() {
-      const skillName = this.memberWithProduction.member.pokemon.skill.name
-        .toLowerCase()
-        .replace(/[()]/g, '')
-        .replace(/\s+/g, '-')
+      const skill = this.memberWithProduction.member.pokemon.skill
+      const skillName = this.normalize(skill.name)
+
+      // Remove `-range` from variable Dream Shard Magnet and Charge Strength skills.
+      // Perhaps a future refactor will make `baseSkill` a field for all Mainskills,
+      // and then the variable skills won't need this hack.
+      const baseSkillName = (
+        skill.isModified ? this.normalize((skill as ModifiedMainskill).baseSkill.name) : skillName
+      ).replace('-range', '')
 
       // Dynamically import the component based on the skill name
       return defineAsyncComponent(
         () =>
           import(
-            `@/components/calculator/results/member-results/member-production-header/member-production-skill-details/${skillName.toLowerCase()}-details.vue`
+            `@/components/calculator/results/member-results/member-production-header/member-production-skill-details/${baseSkillName}/${skillName}-details.vue`
           )
       )
+    }
+  },
+  methods: {
+    normalize(name: string) {
+      return name.toLowerCase().replace(/[()]/g, '').replace(/\s+/g, '-')
     }
   }
 })
